Implement new mock class structure for existing unit tests (#21)
Previous approach to mocking relied heavily on using C++ Lambda's to mock certain function
calls in unit tests. This was not an efficient long term approach.

A new class structure has been introduced when a class needs to be mocked for unit tests. An
interface has now been created for the implementation class and mock class to inherit from.
This also makes certain external dependencies as objects of that class.

The changes here apply primarily for the glTF operations and its relationship with ConfigImportExport.
This will serve as a basis for any changes to be made in the future, and accounts for a subset of 
changes that need to be made across Ambit.